all admin methods for songs

+++ src/main/java/com/spotily/app/song/SongService.java
@@ -27,22 +27,32 @@ public class SongService {
         }
     }
 
-    public void deleteSong(int id) {
+    public void deleteSong(int adminId, int id) {
 //        Optional<Song> songOptional = songDataAccessService.deleteSong(id);
 
 //        if(songOptional.isEmpty()){
 //            throw new ResourceNotFound("Song not found, field can't be empty");
 //        }
-        songDataAccessService.deleteSong(id);
+        ArrayList<Integer> AdminIds = songDataAccessService.getAdmin();
+        if (AdminIds.contains(adminId)) {
+            songDataAccessService.deleteSong(id);
+        } else {
+            throw new UnsupportedOperationException("Only Admins are allowed to delete songs");
+        }
     }
 
-    public void editSongDetails(int id, Song song) {
+    public void editSongDetails(int adminId, int id, Song song) {
 
 //        Optional<Song> songOptional = songDataAccessService.editSongDetails(id, song);
 //
 //        if (songOptional.isEmpty()){
 //            throw new ResourceNotFound("Song not found, field can't be empty");
 //        }
-        songDataAccessService.editSongDetails(id, song);
+        ArrayList<Integer> AdminIds = songDataAccessService.getAdmin();
+        if (AdminIds.contains(adminId)) {
+            songDataAccessService.editSongDetails(id, song);
+        } else {
+            throw new UnsupportedOperationException("Only Admins are allowed to delete songs");
+        }
     }
 }
